Add precision-recall curve logging to test evaluation (#47)
* add precision recall logging

+++ dpvt/wrapper.py
@@ -101,7 +101,7 @@ def custom_collate(items):
 class TreeDataset(Dataset):
     def __init__(self, data, labels):
         self.data = data
-        self.labels = self.add_padding(labels).to(torch.float64)
+        self.labels = self.add_padding(labels).to(torch.float32)
         self.mask = self.mask_pendant_edges(data)
 
     def __len__(self):
@@ -230,12 +230,12 @@ class TraversalDataset(Dataset):
         mutations = torch.full(
             (len(trees), max_n_nodes, max_n_sites, 4),
             -1,
-            dtype=torch.float64,
+            dtype=torch.float32,
             device="cpu",
         )
         # from actual 0 entries representing no mutation
         traversal = torch.full(
-            (len(trees), 2, max_n_int_nodes, 3), -1, dtype=torch.float64, device="cpu"
+            (len(trees), 2, max_n_int_nodes, 3), -1, dtype=torch.float32, device="cpu"
         )
         timings["  allocate_tensors"] = time.perf_counter() - t0
         tensor_size_gb = (mutations.numel() + traversal.numel()) * 4 / (1024**3)
@@ -345,7 +345,7 @@ class TraversalDataset(Dataset):
     def pad_labels(self, labels):
         print(f"Padding labels for {len(labels)} trees...")
         label_tensors = [
-            torch.tensor(label, dtype=torch.float64, device="cpu") for label in labels
+            torch.tensor(label, dtype=torch.float32, device="cpu") for label in labels
         ]
         result = pad_sequence(label_tensors, batch_first=True, padding_value=0)
         print("  Label padding complete!")
